moved allowed_hosts to a submodule
