+++ service-providers/network-requester/src/lib.rs
@@ -1,11 +1,11 @@
 // Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod allowed_hosts;
 pub mod config;
 pub mod core;
 pub mod error;
 mod reply;
+pub mod request_filter;
 mod socks5;
 mod statistics;
