+++ service-providers/network-requester/src/request_filter/mod.rs
@@ -1,0 +1,9 @@
+// Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
+
+/// Old request filtering based on the allowed.list files.
+pub mod allowed_hosts;
+
+pub trait RequestFilter {
+    //
+}
